Renamed text generation function file
