+++ src/utils/dealTooltipText.ts
